Updated frontend and backend

+++ endpoints/jobs.py
@@ -1,12 +1,12 @@
 ## Different Endpoints for Job(s)
 from datetime import datetime
-from fastapi import Depends, HTTPException
+from fastapi import Depends, HTTPException, Request
 from sqlalchemy.orm import Session
 from typing import List
 
 from database.db import get_db
 from database.schema import Job_Scheme, Job_Create
-from database.model import Job
+from database.model import Job, User
 from Tracker import app
 
 @app.get("/api/jobs", response_model=List[Job_Scheme])
@@ -21,7 +21,7 @@ def get_jobs_list(db: Session = Depends(get_db)):
     return jobs
 
 @app.post("/api/job", response_model=Job_Scheme)
-def create_job(job: Job_Create, db: Session = Depends(get_db)):
+def create_job(job: Job_Create, req: Request, db: Session = Depends(get_db)):
     '''
     Creates a new job
 
@@ -31,7 +31,18 @@ def create_job(job: Job_Create, db: Session = Depends(get_db)):
         200 - data of the Job created. Data in Job_Scheme schema format
         401 - when  the user who creates the job is a Candidate and not a recruiter
     '''
-    job_item = Job(**job.dict())
+    auth_header = req.headers['Authorization'].split()
+    if auth_header[0] == "user":
+        req_user = auth_header[1]
+    else:
+        raise HTTPException(status_code=401, detail="Invalid Authorization Header !")
+    job_dict = job.dict()
+    if job_dict['creator'] != req_user:
+        raise HTTPException(status_code=401, detail="User not permitted.. Job Creator & Auth header are pointing different person")
+    user = db.query(User).filter(User.email == job_dict['creator']).first()
+    if not user or not user.is_recruiter:
+        raise HTTPException(status_code=401, detail="User not permitted")
+    job_item = Job(**job_dict)
     db.add(job_item)
     db.commit()
     db.refresh(job_item)
@@ -54,7 +65,7 @@ def get_job(job_id: int, db: Session = Depends(get_db)):
     return job
 
 @app.delete("/api/job/{job_id}")
-def delete_job(job_id: int, db: Session = Depends(get_db)):
+def delete_job(job_id: int, req: Request, db: Session = Depends(get_db)):
     '''
     Delete the job from the DB corresponding to the Job id provided.
 
@@ -63,8 +74,17 @@ def delete_job(job_id: int, db: Session = Depends(get_db)):
     Response:
         200 - Json message stating the success
         404 - When the job is not found (Json Format)
+        401 - when the user is a candidate, not supposed to delete the job
     '''
-    job = db.query(Job).filter_by(Job.id == job_id)
+    auth_header = req.headers['Authorization'].split()
+    if auth_header[0] == "user":
+        req_user = auth_header[1]
+    else:
+        raise HTTPException(status_code=401, detail="Invalid Authorization Header !")
+    user = db.query(User).filter(User.email == req_user).first()
+    if not user or not user.is_recruiter:
+        raise HTTPException(status_code=401, detail="User not permitted")
+    job = db.query(Job).filter(Job.id == job_id).first()
     if not job:
         raise HTTPException(status_code=404, detail="Job not found")
     db.delete(job)
@@ -72,7 +92,7 @@ def delete_job(job_id: int, db: Session = Depends(get_db)):
     return {'msg': "Deletion successful"}
 
 @app.put("/api/job/{job_id}")
-def apply_job(job_id: int, db: Session = Depends(get_db)):
+def apply_job(job_id: int, req: Request, db: Session = Depends(get_db)):
     '''
     Apply to the job whose Job id is provided
 
@@ -84,8 +104,15 @@ def apply_job(job_id: int, db: Session = Depends(get_db)):
         400 - when the user has already applied for the job
         401 - when the user is a recruiter who is not supposed to apply
     '''
-
-    job = db.query(Job).filter_by(Job.id == job_id)
+    auth_header = req.headers['Authorization'].split()
+    if auth_header[0] == "user":
+        req_user = auth_header[1]
+    else:
+        raise HTTPException(status_code=401, detail="Invalid Authorization Header !")
+    user = db.query(User).filter(User.email == req_user).first()
+    if not user or user.is_recruiter:
+        raise HTTPException(status_code=401, detail="User not permitted")
+    job = db.query(Job).filter(Job.id == job_id).first()
     if not job:
         raise HTTPException(status_code=404, detail="Job not found")
     if job.is_applied:
